fix package naming
