+++ src/main/java/com/rocs/infirmary/desktop/InfirmarySystemApplication.java
@@ -4,10 +4,10 @@ import com.rocs.infirmary.desktop.app.facade.dashboard.DashboardFacade;
 import com.rocs.infirmary.desktop.app.facade.dashboard.impl.DashboardFacadeImpl;
 import com.rocs.infirmary.desktop.data.model.person.student.Student;
 import com.rocs.infirmary.desktop.data.model.person.Person;
-import com.rocs.infirmary.desktop.data.model.reports.ailment.CommonAilmentsReport;
-import com.rocs.infirmary.desktop.data.model.reports.lowstock.LowStockReport;
-import com.rocs.infirmary.desktop.data.model.reports.visit.FrequentVisitReport;
-import com.rocs.infirmary.desktop.data.model.reports.medication.MedicationTrendReport;
+import com.rocs.infirmary.desktop.data.model.report.ailment.CommonAilmentsReport;
+import com.rocs.infirmary.desktop.data.model.report.lowstock.LowStockReport;
+import com.rocs.infirmary.desktop.data.model.report.visit.FrequentVisitReport;
+import com.rocs.infirmary.desktop.data.model.report.medication.MedicationTrendReport;
 import com.rocs.infirmary.desktop.app.facade.student.record.impl.StudentMedicalRecordFacadeImpl;
 
 import java.text.ParseException;
