+++ src/main/java/com/rocs/infirmary/desktop/app/facade/dashboard/impl/DashboardFacadeImpl.java
@@ -3,10 +3,10 @@ package com.rocs.infirmary.desktop.app.facade.dashboard.impl;
 import com.rocs.infirmary.desktop.app.facade.dashboard.DashboardFacade;
 import com.rocs.infirmary.desktop.data.dao.report.dashboard.DashboardDao;
 import com.rocs.infirmary.desktop.data.dao.report.dashboard.impl.DashboardDaoImpl;
-import com.rocs.infirmary.desktop.data.model.reports.ailment.CommonAilmentsReport;
-import com.rocs.infirmary.desktop.data.model.reports.lowstock.LowStockReport;
-import com.rocs.infirmary.desktop.data.model.reports.visit.FrequentVisitReport;
-import com.rocs.infirmary.desktop.data.model.reports.medication.MedicationTrendReport;
+import com.rocs.infirmary.desktop.data.model.report.ailment.CommonAilmentsReport;
+import com.rocs.infirmary.desktop.data.model.report.lowstock.LowStockReport;
+import com.rocs.infirmary.desktop.data.model.report.visit.FrequentVisitReport;
+import com.rocs.infirmary.desktop.data.model.report.medication.MedicationTrendReport;
 
 import java.util.Date;
 import java.util.List;
