+++ src/main/java/com/rocs/infirmary/desktop/data/dao/report/dashboard/impl/DashboardDaoImpl.java
@@ -3,11 +3,10 @@ package com.rocs.infirmary.desktop.data.dao.report.dashboard.impl;
 import com.rocs.infirmary.desktop.data.connection.ConnectionHelper;
 import com.rocs.infirmary.desktop.data.dao.report.dashboard.DashboardDao;
 import com.rocs.infirmary.desktop.data.model.person.Person;
-import com.rocs.infirmary.desktop.data.model.person.student.Student;
-import com.rocs.infirmary.desktop.data.model.reports.ailment.CommonAilmentsReport;
-import com.rocs.infirmary.desktop.data.model.reports.lowstock.LowStockReport;
-import com.rocs.infirmary.desktop.data.model.reports.visit.FrequentVisitReport;
-import com.rocs.infirmary.desktop.data.model.reports.medication.MedicationTrendReport;
+import com.rocs.infirmary.desktop.data.model.report.ailment.CommonAilmentsReport;
+import com.rocs.infirmary.desktop.data.model.report.lowstock.LowStockReport;
+import com.rocs.infirmary.desktop.data.model.report.visit.FrequentVisitReport;
+import com.rocs.infirmary.desktop.data.model.report.medication.MedicationTrendReport;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
